WIP Add initial build config for nRF52840 Dongle
TODO Finish cleaning up variant.h.

This adds a board and variant for the nRF52840 Dongle. It includes
a PlatformIO environment that builds a blinky app, and a shell
script to package the app and flash it to the device.

The nrfutil used by PlatformIO does not work correctly with the Open
Bootloader, therefore the suboptimal approach to using a separate shell
script for uploading was necessary.

The PlatformIO environment requires SoftDevice 7.3.0 to be present on
the device. The blinky app does not use the bluetooth stack, but because
getting the correct SoftDevice on the chip turned out to be difficult,
a main objective of this commit is to document the process for configuring
the correct SoftDevice and demonstrate its correct configuration.

The SoftDevice bluetooth stack should be ready to use given the work in
this commit.